Create Welcome Message
Appusamy is running a hotel called Brindavan and he wants to greet the customers with the message "Welcome to Brindavan". 
Help Appusamy by completing the program to print the desired welcome message. 
Please note that the output is case sensitive. 
So print W and B in upper case. 
Output: Welcome to Brindavan