Create Sum of ages of Husband and wife
Honeymoon hotel in Simla awards a prize to the youngest couple staying in their hotel. For this they wish to write a program to calculate the sum of the ages of husband and wife. Help the hotel by filling in the program code below. Input Format: The first line denotes the age of the husband. The second line denotes the age of the wife 
Example Input/Output: 
Input: 25 21 
Output: 
46