Create Sum of ODD digits in a Number
A number N is passed as the input. The program must print the sum of odd digits in the number N. Input Format: The first line denotes the value of N. Output Format: The sum of odd digits in number N Boundary Conditions: 1 <= N <= 9999999 
Example Input/Output 1: Input: 4165 Output: 6 
Explanation: The odd digits are 1 ,5 and their sum is 6. 
Example Input/Output 2: Input: 224 Output: 0