Create MFIB - Convert String to Upper Case
Fill in the blanks with code so that the program must accept a string and and convert the characters in the string to upper case. Then the program must print the modified string as the output. Example Input/Output 1: Input: SkillRack Output: SKILLRACK Example Input/Output 2: Input: flowers Output: FLOWERS