Create Split string
Fill in the blanks with code so that the program accepts a string and splits the string using a delimiter ':'. Then the program must print the values in the string array obtained by splitting the string as the output. Example Input/Output 1: Input: one:two:three:four Output: one two three four