Create MFIB - Convert String to char Array
Fill in the blanks with code so that the program accepts a string and converts the string into a character array. Then the program must print each character in the character array as the output. 
Example Input/Output 1: Input: skillrack Output: s k i l l r a c k